requirements.txt
pillow
pandas
scikit-learn
numpy
tqdm
pytorch
